Extract shared copyToClipboard helper, dedupe ShareButton + JSON copy
Both the new JSON copy button and ShareButton had their own clipboard +
execCommand fallback; centralize in src/lib/clipboard.ts.
